requirements.txt
python-dotenv==1.0.1
termcolor==2.4.0
chardet==5.2.0
pydantic==2.9.2
openai==1.63.2
instructor==1.5.0
bs4==0.0.2
requests==2.32.3
scapy
pyshark
sentence-transformers
duckduckgo-search
pandas
numpy
matplotlib
SciencePlots